requirements.txt
feedparser>=6.0.11
requests>=2.31.0
google-generativeai>=0.8.0
openai>=1.50.0
python-dotenv>=1.0.1
